requirements.txt
gpytorch==1.3.1
torch==1.8.1
ax_platform==0.1.19
botorch==0.3.3
pyro_ppl==1.6.0
